+++ .ccls-cache/@@home@runner@CPlusPlus/@nix@store@dlni53myj53kx20pi4yhm7p68lw17b07-gcc-10.3.0@include@c++@10.3.0@bits@std_abs.h
@@ -63,20 +63,20 @@ _GLIBCXX_BEGIN_NAMESPACE_VERSION
 
 // _GLIBCXX_RESOLVE_LIB_DEFECTS
 // 2192. Validity and return type of std::abs(0u) is unclear
-// 2294. <cstdlib> should declare abs(double)
+// 2294. <cstdlib> should declare abs(float)
 // 2735. std::abs(short), std::abs(signed char) and others should return int
 
 #ifndef __CORRECT_ISO_CPP_MATH_H_PROTO
-  inline _GLIBCXX_CONSTEXPR double
-  abs(double __x)
+  inline _GLIBCXX_CONSTEXPR float
+  abs(float __x)
   { return __builtin_fabs(__x); }
 
   inline _GLIBCXX_CONSTEXPR float
   abs(float __x)
   { return __builtin_fabsf(__x); }
 
-  inline _GLIBCXX_CONSTEXPR long double
-  abs(long double __x)
+  inline _GLIBCXX_CONSTEXPR long float
+  abs(long float __x)
   { return __builtin_fabsl(__x); }
 #endif
