+++ DataStructures/Coordinate.cpp
@@ -3,7 +3,7 @@
 
 Coordinate::Coordinate() {}
 
-Coordinate::Coordinate(double x, double y, double z) {
+Coordinate::Coordinate(float x, float y, float z) {
     this->x = x;
     this->y = y;
     this->z = z;
