+++ DataStructures/Vector.h
@@ -7,27 +7,27 @@
 class Vector3D {
 public:
   Vector3D();
-  Vector3D(double x, double y, double z);
+  Vector3D(float x, float y, float z);
   Vector3D(Coordinate cord);
-  double getLength();
-  double getElementAt(int index);
-  bool setLength(double);
-  double *getVector();
-  bool setVector(double x, double y, double z);
+  float getLength();
+  float getElementAt(int index);
+  bool setLength(float);
+  float *getVector();
+  bool setVector(float x, float y, float z);
   bool setVector(Coordinate cord);
-  double calcLength();
+  float calcLength();
   void normalize();
   friend std::ostream& operator<<(std::ostream& os, Vector3D& a);
-  static double dot(Vector3D a, Vector3D b) {
-    double *pa = a.getVector();
-    double *pb = b.getVector();
+  static float dot(Vector3D a, Vector3D b) {
+    float *pa = a.getVector();
+    float *pb = b.getVector();
 
     return pa[0] * pb[0] + pa[1] * pb[1] + pa[2] * pb[2];
   }
   static Vector3D cross(Vector3D a, Vector3D b) {
-    double cx = (a.getElementAt(1)*b.getElementAt(2)) - (a.getElementAt(2)*b.getElementAt(1));
-    double cy = (a.getElementAt(2)*b.getElementAt(0)) - (a.getElementAt(0)*b.getElementAt(2));
-    double cz = (a.getElementAt(0)*b.getElementAt(1)) - (a.getElementAt(1)*b.getElementAt(0));
+    float cx = (a.getElementAt(1)*b.getElementAt(2)) - (a.getElementAt(2)*b.getElementAt(1));
+    float cy = (a.getElementAt(2)*b.getElementAt(0)) - (a.getElementAt(0)*b.getElementAt(2));
+    float cz = (a.getElementAt(0)*b.getElementAt(1)) - (a.getElementAt(1)*b.getElementAt(0));
     return Vector3D(cx,cy,cz);
   }
 
@@ -49,14 +49,14 @@ public:
     return false;
   }
 
-  Vector3D operator*(double right) {
+  Vector3D operator*(float right) {
     return Vector3D(vector[0] * right, vector[1] * right, vector[2] * right);
   }
-  Vector3D operator/(double right) {
+  Vector3D operator/(float right) {
     Vector3D v = Vector3D(vector[0] / right, vector[1] / right, vector[2] / right);
     return v;
   }
-  Vector3D operator+(double right){
+  Vector3D operator+(float right){
     Vector3D v = Vector3D(vector[0] + right, vector[1] + right, vector[2] + right);
     return v;
   }
@@ -79,36 +79,36 @@ public:
   }
 
 private:
-  double vector[3];
-  double lenght;
+  float vector[3];
+  float lenght;
   int maxIndex;
 };
 
 class Vector4D {
 public:
   Vector4D();
-  Vector4D(double x, double y, double z);
+  Vector4D(float x, float y, float z);
   Vector4D(Coordinate cord);
   Vector4D(Vector3D v3d);
   Coordinate toCoord();
-  double getLength();
-  double getElementAt(int index);
-  bool setLength(double);
-  double *getVector();
-  bool setVector(double x, double y, double z,double h);
+  float getLength();
+  float getElementAt(int index);
+  bool setLength(float);
+  float *getVector();
+  bool setVector(float x, float y, float z,float h);
   bool setVector(Coordinate cord);
-  double calcLength();
+  float calcLength();
   void normalize();
   friend std::ostream& operator<<(std::ostream& os, Vector4D& a);
   static Vector4D cross(Vector4D a, Vector4D b) {
-    double cx = (a.getElementAt(1)*b.getElementAt(2)) - (a.getElementAt(2)*b.getElementAt(1));
-    double cy = (a.getElementAt(2)*b.getElementAt(0)) - (a.getElementAt(0)*b.getElementAt(2));
-    double cz = (a.getElementAt(0)*b.getElementAt(1)) - (a.getElementAt(1)*b.getElementAt(0));
+    float cx = (a.getElementAt(1)*b.getElementAt(2)) - (a.getElementAt(2)*b.getElementAt(1));
+    float cy = (a.getElementAt(2)*b.getElementAt(0)) - (a.getElementAt(0)*b.getElementAt(2));
+    float cz = (a.getElementAt(0)*b.getElementAt(1)) - (a.getElementAt(1)*b.getElementAt(0));
     return Vector4D(cx,cy,cz);
   }
-  static double dot(Vector4D a, Vector4D b) {
-    double *pa = a.getVector();
-    double *pb = b.getVector();
+  static float dot(Vector4D a, Vector4D b) {
+    float *pa = a.getVector();
+    float *pb = b.getVector();
 
     return pa[0] * pb[0] + pa[1] * pb[1] + pa[2] * pb[2];
   }
@@ -131,14 +131,14 @@ public:
     return false;
   }
 
-  Vector4D operator*(double right) {
+  Vector4D operator*(float right) {
     return Vector4D(vector[0] * right, vector[1] * right, vector[2] * right);
   }
-  Vector4D operator/(double right) {
+  Vector4D operator/(float right) {
     Vector4D v = Vector4D(vector[0] / right, vector[1] / right, vector[2] / right);
     return v;
   }
-  Vector4D operator+(double right){
+  Vector4D operator+(float right){
     Vector4D v = Vector4D(vector[0] + right, vector[1] + right, vector[2] + right);
     return v;
   }
@@ -159,8 +159,8 @@ public:
   }
 
 private:
-  double vector[4];
-  double lenght;
+  float vector[4];
+  float lenght;
   int maxIndex;
 };
