+++ World/Space3D.cpp
@@ -12,4 +12,4 @@
 #include <stdio.h>
 
 
-Coordinate canvasToViewport(double x, double y){return Coordinate{0,0,0};};
+Coordinate canvasToViewport(float x, float y){return Coordinate{0,0,0};};
